feat: local models off by default + collapsible settings block (#346)
De-emphasise vLLM / Ollama / LM Studio:
- New collapsed-by-default "Локальные модели (экспериментально)" block built
  on the existing accessible SettingsSection (keyboard + ARIA).
- vLLM/Ollama status grid and LM Studio panel moved under it; cloud providers
  stay prominent. Saved endpoints are never deleted, only hidden/inactive.
- Backend stays endpoint-gated (off on a clean install); add regression tests.
- Extract pure local-provider logic into client/src/lib/local-providers.ts
  with unit tests (collapsed default, off-by-default, enable restores).
